+++ Codes/HotelReserveSystem_Client/src/main/java/bl_Stub/hotelblservice_Stub/CheckOrderedHotelServiceImpl_Stub.java
@@ -11,10 +11,10 @@ public class CheckOrderedHotelServiceImpl_Stub implements CheckOrderedHotelServi
 	public String businessDistrict;
 	public String hotelAddress;
 	public int starLevel;
-	public int mark;
+	public float mark;
 	
 	public CheckOrderedHotelServiceImpl_Stub(String hotelName, String businessDistrict, String hotelAddress, int starLevel,
-			int mark) {
+			float mark) {
 		this.hotelName = hotelName;
 		this.businessDistrict = businessDistrict;
 		this.hotelAddress = hotelAddress;
